Working on ready plans 0.0.6
Complete with UI about 'view ready plan'

Little refactor. Need test on bug.
