+++ lib/features/auth/auth_reset_password/presentation/next_step_after_enter_new_passoword.dart
@@ -86,6 +86,7 @@ class _NewNextStepAfterEnterNewPasswordState
                   }
                 : () {},
             style: ButtonStyle(
+                elevation: const WidgetStatePropertyAll(0),
                 fixedSize: WidgetStatePropertyAll(
                     Size(MediaQuery.of(context).size.width, 60)),
                 backgroundColor:
