+++ lib/features/auth/auth_reset_password/presentation/next_step_after_send_recovery_code.dart
@@ -82,6 +82,7 @@ class _NewNextStepAfterSendRecoveryCodeButtonState
                     showError(false);
                   },
             style: ButtonStyle(
+                elevation: const WidgetStatePropertyAll(0),
                 fixedSize: WidgetStatePropertyAll(
                     Size(MediaQuery.of(context).size.width, 60)),
                 backgroundColor:
