+++ lib/features/auth/auth_sign_in/presentation/sign_in_button.dart
@@ -88,6 +88,7 @@ class _NewSignInButtonState extends ConsumerState<SignInButton> {
                     }
                   },
             style: ButtonStyle(
+                elevation: const WidgetStatePropertyAll(0),
                 fixedSize: WidgetStatePropertyAll(
                     Size(MediaQuery.of(context).size.width, 60)),
                 backgroundColor:
