+++ lib/features/auth/auth_sign_up/presentation/components/next_step_after_weight_age_select_button.dart
@@ -37,6 +37,7 @@ class NextStepAfterWeightAgeSelectButton extends ConsumerWidget {
               context.goNamed(weightOrAge ? 'height' : 'goal');
             },
             style: ButtonStyle(
+                elevation: const WidgetStatePropertyAll(0),
                 fixedSize: WidgetStatePropertyAll(
                     Size(MediaQuery.of(context).size.width, 60)),
                 backgroundColor:
