+++ lib/features/auth/auth_sign_up/presentation/gender/next_step_after_gender_button.dart
@@ -40,6 +40,7 @@ class NextStepAfterGenderButton extends ConsumerWidget {
                     context.goNamed('weight');
                   },
             style: ButtonStyle(
+                elevation: const WidgetStatePropertyAll(0),
                 fixedSize: WidgetStatePropertyAll(
                     Size(MediaQuery.of(context).size.width, 60)),
                 backgroundColor:
