+++ lib/features/auth/auth_sign_up/presentation/goal/next_step_after_goal_select_button.dart
@@ -33,6 +33,7 @@ class NextStepAfterGoalSelectButton extends ConsumerWidget {
               context.goNamed('level');
             },
             style: ButtonStyle(
+                elevation: const WidgetStatePropertyAll(0),
                 fixedSize: WidgetStatePropertyAll(
                     Size(MediaQuery.of(context).size.width, 60)),
                 backgroundColor:
