+++ lib/features/auth/auth_sign_up/presentation/height/next_step_after_height_select_button.dart
@@ -33,6 +33,7 @@ class NextStepAfterHeightSelectButton extends ConsumerWidget {
               context.goNamed('age');
             },
             style: ButtonStyle(
+                elevation: const WidgetStatePropertyAll(0),
                 fixedSize: WidgetStatePropertyAll(
                     Size(MediaQuery.of(context).size.width, 60)),
                 backgroundColor:
