+++ lib/features/auth/auth_sign_up/presentation/level/next_step_after_level.dart
@@ -40,6 +40,7 @@ class NextStepAfterLevelSelectButton extends ConsumerWidget {
                     context.goNamed('signup');
                   },
             style: ButtonStyle(
+                elevation: const WidgetStatePropertyAll(0),
                 fixedSize: WidgetStatePropertyAll(
                     Size(MediaQuery.of(context).size.width, 60)),
                 backgroundColor:
