+++ lib/features/auth/auth_sign_up/presentation/sign_up/sign_up_button.dart
@@ -90,6 +90,7 @@ class _NewSignUpButtonState extends ConsumerState<SignUpButton> {
               }
             },
             style: ButtonStyle(
+                elevation: const WidgetStatePropertyAll(0),
                 fixedSize: WidgetStatePropertyAll(
                     Size(MediaQuery.of(context).size.width, 70)),
                 backgroundColor:
