+++ lib/features/home/future_profile/presentation/height_changer/next_step_after_height_home_button.dart
@@ -51,6 +51,7 @@ class _NextStepAfterHeightHomeButtonState
                   }
                 : () {},
             style: ButtonStyle(
+                elevation: const WidgetStatePropertyAll(0),
                 fixedSize: WidgetStatePropertyAll(
                     Size(MediaQuery.of(context).size.width, 70)),
                 backgroundColor: WidgetStatePropertyAll(
