+++ lib/features/home/future_profile/presentation/weight_age_changer/next_step_after_age_weight_home_button.dart
@@ -59,6 +59,7 @@ class _NextStepAfterAgeHeightHomeButtonState
                   }
                 : () {},
             style: ButtonStyle(
+                elevation: const WidgetStatePropertyAll(0),
                 fixedSize: WidgetStatePropertyAll(
                     Size(MediaQuery.of(context).size.width, 70)),
                 backgroundColor: WidgetStatePropertyAll(
