+++ lib/features/train/create_training_plan/domain/controllers/confrim_ready_plan_controller.dart
@@ -1,6 +1,5 @@
 import 'package:fitflow/features/general_comonents/exercise_model.dart';
 import 'package:fitflow/features/train/create_training_plan/data/providers/get_ready_plans_data_provider.dart';
-import 'package:fitflow/features/train/create_training_plan/domain/models/ready_training_plan_model.dart';
 import 'package:fitflow/features/train/get_training_plan/domain/providers/get_training_plan_domain_provider.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
