+++ lib/features/train/create_training_plan/presentation/select_way_of_creating_train_plan/components/confrim_way_of_creating_button.dart
@@ -38,6 +38,7 @@ class ConfrimWayOfCreatingButton extends ConsumerWidget {
                     ref.invalidate(getReadyTrainPlanDomainProvider);
                   },
             style: ButtonStyle(
+                elevation: const WidgetStatePropertyAll(0),
                 fixedSize: WidgetStatePropertyAll(
                     Size(MediaQuery.of(context).size.width, 60)),
                 backgroundColor:
