+++ lib/features/train/get_temp_week_and_today_train_progress/presentation/components/buttons_today_day/chill_day/components/chill_day_text.dart
@@ -13,6 +13,7 @@ class ChillDayTextInsideButton extends StatelessWidget {
       child: TextButton(
         onPressed: () {},
         style: const ButtonStyle(
+            elevation: WidgetStatePropertyAll(0),
             overlayColor: WidgetStatePropertyAll(Colors.transparent),
             padding: WidgetStatePropertyAll(EdgeInsets.all(0))),
         child: ShaderMask(
