+++ lib/features/train/get_temp_week_and_today_train_progress/presentation/components/buttons_today_day/empty_plan/components/create_plan_text.dart
@@ -16,6 +16,7 @@ class CreatePlanTextInsideButton extends StatelessWidget {
           context.goNamed('newtrainplan');
         },
         style: const ButtonStyle(
+            elevation: WidgetStatePropertyAll(0),
             overlayColor: WidgetStatePropertyAll(Colors.transparent),
             padding: WidgetStatePropertyAll(EdgeInsets.all(0))),
         child: SizedBox(
