+++ lib/features/train/get_temp_week_and_today_train_progress/presentation/components/buttons_today_day/start_train/components/start_train_text.dart
@@ -11,6 +11,7 @@ class StartTrainTextIndiseButton extends StatelessWidget {
     return TextButton(
       onPressed: () {},
       style: const ButtonStyle(
+          elevation: WidgetStatePropertyAll(0),
           overlayColor: WidgetStatePropertyAll(Colors.transparent),
           padding: WidgetStatePropertyAll(EdgeInsets.all(0))),
       child: ShaderMask(
