+++ lib/features/train/get_temp_week_and_today_train_progress/presentation/components/buttons_today_day/view_progress/components/view_progress_text.dart
@@ -16,6 +16,7 @@ class ViewProgressTextInsideButton extends StatelessWidget {
           context.goNamed('/progresshome');
         },
         style: const ButtonStyle(
+            elevation: WidgetStatePropertyAll(0),
             overlayColor: WidgetStatePropertyAll(Colors.transparent),
             padding: WidgetStatePropertyAll(EdgeInsets.all(0))),
         child: Text('Отследить прогресс',
